move extension to AsakusaSatellite.Message to a separated file
