+++ AsakusaSatellite/RoomViewController.swift
@@ -13,20 +13,6 @@ import AsakusaSatellite
 private let kCellID = "Cell"
 
 
-extension Message {
-    var hasHTML: Bool {return body != htmlBody}
-    var html: String { return "<!DOCTYPE html>"
-        + "<html>"
-        + "<head>"
-        + "<meta content=\"width=device-width, initial-scale=1.0, maximum-scale=4.0, user-scalable=yes\" name=\"viewport\">"
-        + "<link href=\"assets/application.css\" media=\"screen\" rel=\"stylesheet\" type=\"text/css\">"
-        + "<style>body{margin:0; padding:0; background-color: white;} iframe{margin-left:-5px;}</style>"
-        + "</head>"
-        + "<body><div id=\"AsakusaSatMessageContent\" style=\"padding: 8px;\">\(htmlBody)</div></body>"
-        + "</html>" }
-}
-
-
 class RoomViewController: UIViewController, UITableViewDataSource, UITableViewDelegate, UIScrollViewDelegate, UIWebViewDelegate {
     let client: Client
     var pusher: MessagePusherClient?
